+++ src/app/shared/directives/color-click/color-click.directive.spec.ts
@@ -1,0 +1,8 @@
+import { ColorClickDirective } from './color-click.directive';
+
+describe('ColorClickDirective', () => {
+  it('should create an instance', () => {
+    const directive = new ColorClickDirective();
+    expect(directive).toBeTruthy();
+  });
+});